Scripts rework part 2

+++ datafari-mcf/bin/mcf_crawler_agent.sh
@@ -15,9 +15,8 @@ OPTIONS=$(cat "$OPTIONSFILE")
 cmd_start() {
     echo "Starting MCF Agent ..."
     ./executecommand.sh org.apache.manifoldcf.core.LockClean
-    start-stop-daemon --background --chdir=$MCF_HOME --start --make-pidfile --pidfile $MCF_PID_FILE --exec \
-    "$JAVA_HOME/bin/java" -- $OPTIONS org.apache.manifoldcf.agents.AgentRun
-    sleep 1
+    nohup "$JAVA_HOME/bin/java" $OPTIONS org.apache.manifoldcf.agents.AgentRun &
+    echo $! > $MCF_PID_FILE
 	echo "MCF Agent started with PID $(cat $MCF_PID_FILE)"
     return 0
 }
@@ -25,7 +24,15 @@ cmd_start() {
 cmd_stop() {
     echo -n "Stopping MCF Agent ..."
     ./executecommand.sh org.apache.manifoldcf.agents.AgentStop
-    rm -f $MCF_PID_FILE
+    #Wait few seconds to let the MCF agent stopping by itself
+    sleep 10
+    mcf_pid=$(cat $MCF_PID_FILE)
+    if ps -p $mcf_pid > /dev/null
+    then
+       echo "Could not stop the MCF crawler agent process ! It's PID file will not be removed !"
+    else
+      rm -f $MCF_PID_FILE
+    fi
 }
 
 cmd_status() {
